backend/requirements_utf8.txt
# Web 框架和服务器
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2

# 数据库
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1

# AI 和机器学习
openai>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# 工具和辅助库
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.1
python-jose>=3.3.0
passlib>=1.7.4

# 开发工具
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910
